fixed issues with routes

+++ portfolio/router.js
@@ -29,7 +29,8 @@ router.post('/:id', (req, res) => {
   const id = req.params.id;
   const { holdings } = req.body;
 
-  CryptoPortfolio.findOne({ id }, (err, existingCoin) => {
+
+  CryptoPortfolio.findOne({ _creator: req.user._id, id }, (err, existingCoin) => {
     if (existingCoin === null) {
       getCoins(id)
         .then(x => x[0])
@@ -46,9 +47,6 @@ router.post('/:id', (req, res) => {
           console.error(err);
           res.status(500).json({ message: 'Internal server error' });
         });
-    } else {
-      const capitalizedId = id.charAt(0).toUpperCase() + id.slice(1);
-      res.json(`${capitalizedId} already in watchlist`);
     }
   });
 });
@@ -57,7 +55,7 @@ router.post('/:id', (req, res) => {
 router.delete('/:id', (req, res) => {
   const id = req.params.id;
 
-  CryptoPortfolio.findOneAndRemove({ id })
+  CryptoPortfolio.findOneAndRemove({ _creator: req.user._id, id })
     .then(count => {
       if (count) {
         res.status(204).end();
